DROID-2023 fixes
