+++ app/src/main/java/com/anytypeio/anytype/ui/main/MainActivity.kt
@@ -15,10 +15,13 @@ import androidx.lifecycle.repeatOnLifecycle
 import androidx.navigation.findNavController
 import com.anytypeio.anytype.BuildConfig
 import com.anytypeio.anytype.R
+import com.anytypeio.anytype.app.AnytypeNotificationService
+import com.anytypeio.anytype.app.AnytypeNotificationService.Companion.NOTIFICATION_TYPE
 import com.anytypeio.anytype.app.DefaultAppActionManager
-import com.anytypeio.anytype.app.NotificationActionInterceptor
+import com.anytypeio.anytype.core_models.Relations
 import com.anytypeio.anytype.core_models.ThemeMode
 import com.anytypeio.anytype.core_models.Wallpaper
+import com.anytypeio.anytype.core_models.primitives.SpaceId
 import com.anytypeio.anytype.core_utils.ext.Mimetype
 import com.anytypeio.anytype.core_utils.ext.parseActionSendMultipleUris
 import com.anytypeio.anytype.core_utils.ext.parseActionSendUri
@@ -35,6 +38,7 @@ import com.anytypeio.anytype.presentation.main.MainViewModel
 import com.anytypeio.anytype.presentation.main.MainViewModel.Command
 import com.anytypeio.anytype.presentation.main.MainViewModelFactory
 import com.anytypeio.anytype.presentation.navigation.AppNavigation
+import com.anytypeio.anytype.presentation.notifications.NotificationAction
 import com.anytypeio.anytype.presentation.wallpaper.WallpaperColor
 import com.anytypeio.anytype.ui.editor.CreateObjectFragment
 import com.anytypeio.anytype.ui.notifications.NotificationsFragment
@@ -160,10 +164,6 @@ class MainActivity : AppCompatActivity(R.layout.activity_main), AppNavigation.Pr
                 proceedWithShareIntent(intent)
             }
         }
-
-        NotificationActionInterceptor.onIntercepted = { action ->
-            vm.onInterceptNotificationAction(action)
-        }
     }
 
     private fun startAppUpdater() {
@@ -215,6 +215,9 @@ class MainActivity : AppCompatActivity(R.layout.activity_main), AppNavigation.Pr
                 }
                 Intent.ACTION_SEND_MULTIPLE -> {
                     proceedWithShareIntent(intent)
+                }
+                AnytypeNotificationService.NOTIFICATION_INTENT_ACTION -> {
+                    proceedWithNotificationIntent(intent)
                 }
             }
         }
@@ -271,6 +274,30 @@ class MainActivity : AppCompatActivity(R.layout.activity_main), AppNavigation.Pr
                 vm.onIntentMultipleImageShare(listOf(uri))
             } else {
                 toast("Could not parse URI")
+            }
+        }
+    }
+
+    private fun proceedWithNotificationIntent(intent: Intent) {
+        when(val type = intent.getIntExtra(NOTIFICATION_TYPE, -1)) {
+            AnytypeNotificationService.REQUEST_TO_JOIN_TYPE -> {
+                val space = intent.getStringExtra(Relations.SPACE_ID)
+                val identity = intent.getStringExtra(Relations.IDENTITY)
+                if (!space.isNullOrEmpty() && !identity.isNullOrEmpty()) {
+                    val notification = intent.getStringExtra(AnytypeNotificationService.NOTIFICATION_ID_KEY).orEmpty()
+                    vm.onInterceptNotificationAction(
+                        action = NotificationAction.Multiplayer.ViewSpaceJoinRequest(
+                            notification = notification,
+                            space = SpaceId(space),
+                            identity = identity
+                        )
+                    )
+                } else {
+                    Timber.w("Missing space or identity")
+                }
+            }
+            else -> {
+                toast("Unknown type: $type")
             }
         }
     }
